+++ src/SettlerIntent.sol
@@ -37,6 +37,7 @@ abstract contract SettlerIntent is MultiCallContext, Permit2PaymentIntent, Settl
     }
 
     address private constant _SENTINEL_SOLVER = 0x0000000000000000000000000000000000000001;
+    address private constant _ADDRESS_MASK = 0xFFfFfFffFFfffFFfFFfFFFFFffFFFffffFfFFFfF;
 
     constructor() {
         assert(_SOLVER_LIST_BASE_SLOT == bytes32(uint256(uint128(uint256(keccak256("SettlerIntentSolverList")) - 1))));
@@ -130,8 +131,8 @@ abstract contract SettlerIntent is MultiCallContext, Permit2PaymentIntent, Settl
         */
         assembly ("memory-safe") {
             // Clean dirty bits.
-            prev := and(0xffffffffffffffffffffffffffffffffffffffff, prev)
-            solver := and(0xffffffffffffffffffffffffffffffffffffffff, solver)
+            prev := and(_ADDRESS_MASK, prev)
+            solver := and(_ADDRESS_MASK, solver)
 
             // A solver of zero is special-cased. It is forbidden to set it because that would
             // corrupt the list.
@@ -141,12 +142,12 @@ abstract contract SettlerIntent is MultiCallContext, Permit2PaymentIntent, Settl
             mstore(0x00, solver)
             mstore(0x20, _SOLVER_LIST_BASE_SLOT)
             let solverSlot := keccak256(0x00, 0x40)
-            let solverSlotValue := and(0xffffffffffffffffffffffffffffffffffffffff, sload(solverSlot))
+            let solverSlotValue := and(_ADDRESS_MASK, sload(solverSlot))
 
             // If the slot contains zero, `addNotRemove` must be true (we are adding a new
             // solver). Likewise if the slot contains nonzero, `addNotRemove` must be false (we are
             // removing one).
-            fail := or(fail, xor(iszero(solverSlotValue), addNotRemove))
+            fail := or(xor(iszero(solverSlotValue), addNotRemove), fail)
 
             // Derive the slot for `prev`.
             mstore(0x00, prev)
@@ -163,7 +164,7 @@ abstract contract SettlerIntent is MultiCallContext, Permit2PaymentIntent, Settl
             // `_SENTINEL_SOLVER`). If we are removing an existing solver, then `prev` must point at
             // `solver.
             fail :=
-                or(fail, xor(and(0xffffffffffffffffffffffffffffffffffffffff, sload(prevSlot)), expectedPrevSlotValue))
+                or(xor(and(_ADDRESS_MASK, sload(prevSlot)), expectedPrevSlotValue), fail)
 
             // Update the linked list. This either points `$[prev]` at `$[solver]` and zeroes
             // `$[solver]` or it points `$[prev]` at `solver` and points `$[solver]` at
@@ -207,10 +208,10 @@ abstract contract SettlerIntent is MultiCallContext, Permit2PaymentIntent, Settl
                 let i := start
                 for {
                     mstore(0x20, _SOLVER_LIST_BASE_SLOT)
-                    let x := and(0xffffffffffffffffffffffffffffffffffffffff, sload(_SOLVER_LIST_START_SLOT))
+                    let x := and(_ADDRESS_MASK, sload(_SOLVER_LIST_START_SLOT))
                 } xor(_SENTINEL_SOLVER, x) {
                     i := add(0x20, i)
-                    x := and(0xffffffffffffffffffffffffffffffffffffffff, sload(keccak256(0x00, 0x40)))
+                    x := and(_ADDRESS_MASK, sload(keccak256(0x00, 0x40)))
                 } {
                     mstore(i, x)
                     mstore(0x00, x)
@@ -240,7 +241,7 @@ abstract contract SettlerIntent is MultiCallContext, Permit2PaymentIntent, Settl
         assembly ("memory-safe") {
             let ptr := mload(0x40)
             mstore(ptr, SLIPPAGE_TYPEHASH)
-            calldatacopy(add(ptr, 0x20), slippage, 0x60)
+            calldatacopy(add(0x20, ptr), slippage, 0x60)
             result := keccak256(ptr, 0x80)
         }
     }